+++ workflows/inngest/src/index.ts
@@ -2,7 +2,7 @@ import { randomUUID } from 'crypto';
 import type { ReadableStream } from 'node:stream/web';
 import { subscribe } from '@inngest/realtime';
 import type { Agent } from '@mastra/core/agent';
-import { AISpanType } from '@mastra/core/ai-tracing';
+import { AISpanType, wrapMastra } from '@mastra/core/ai-tracing';
 import type { TracingContext, AnyAISpan } from '@mastra/core/ai-tracing';
 import { RuntimeContext } from '@mastra/core/di';
 import type { Mastra } from '@mastra/core/mastra';
@@ -563,7 +563,7 @@ export class InngestWorkflow<
           runtimeContext: new RuntimeContext(), // TODO
           resume,
           abortController: new AbortController(),
-          parentSpan: undefined, // TODO: Pass actual parent AI span from workflow execution context
+          currentSpan: undefined, // TODO: Pass actual parent AI span from workflow execution context
         });
 
         return { result, runId };
@@ -768,7 +768,7 @@ export function createStep<
       execute: async ({ inputData, mastra, runtimeContext, tracingContext }) => {
         return params.execute({
           context: inputData,
-          mastra,
+          mastra: wrapMastra(mastra, tracingContext),
           runtimeContext,
           tracingContext,
         });
@@ -879,7 +879,7 @@ export class InngestExecutionEngine extends DefaultExecutionEngine {
     };
     runtimeContext: RuntimeContext;
     abortController: AbortController;
-    parentSpan?: AnyAISpan;
+    currentSpan?: AnyAISpan;
   }): Promise<TOutput> {
     await params.emitter.emit('watch-v2', {
       type: 'start',
@@ -1013,7 +1013,7 @@ export class InngestExecutionEngine extends DefaultExecutionEngine {
   }): Promise<void> {
     let { duration, fn } = entry;
 
-    const sleepSpan = tracingContext?.parentSpan?.createChildSpan({
+    const sleepSpan = tracingContext?.currentSpan?.createChildSpan({
       type: AISpanType.WORKFLOW_SLEEP,
       name: `sleep: ${duration ? `${duration}ms` : 'dynamic'}`,
       attributes: {
@@ -1033,7 +1033,7 @@ export class InngestExecutionEngine extends DefaultExecutionEngine {
           inputData: prevOutput,
           runCount: -1,
           tracingContext: {
-            parentSpan: sleepSpan,
+            currentSpan: sleepSpan,
           },
           getInitData: () => stepResults?.input as any,
           getStepResult: (step: any) => {
@@ -1126,7 +1126,7 @@ export class InngestExecutionEngine extends DefaultExecutionEngine {
   }): Promise<void> {
     let { date, fn } = entry;
 
-    const sleepUntilSpan = tracingContext?.parentSpan?.createChildSpan({
+    const sleepUntilSpan = tracingContext?.currentSpan?.createChildSpan({
       type: AISpanType.WORKFLOW_SLEEP,
       name: `sleepUntil: ${date ? date.toISOString() : 'dynamic'}`,
       attributes: {
@@ -1147,7 +1147,7 @@ export class InngestExecutionEngine extends DefaultExecutionEngine {
           inputData: prevOutput,
           runCount: -1,
           tracingContext: {
-            parentSpan: sleepUntilSpan,
+            currentSpan: sleepUntilSpan,
           },
           getInitData: () => stepResults?.input as any,
           getStepResult: (step: any) => {
@@ -1247,7 +1247,7 @@ export class InngestExecutionEngine extends DefaultExecutionEngine {
     writableStream?: WritableStream<ChunkType>;
     tracingContext?: TracingContext;
   }): Promise<StepResult<any, any, any, any>> {
-    const stepAISpan = tracingContext?.parentSpan?.createChildSpan({
+    const stepAISpan = tracingContext?.currentSpan?.createChildSpan({
       name: `workflow step: '${step.id}'`,
       type: AISpanType.WORKFLOW_STEP,
       input: prevOutput,
@@ -1513,7 +1513,7 @@ export class InngestExecutionEngine extends DefaultExecutionEngine {
           inputData: prevOutput,
           resumeData: resume?.steps[0] === step.id ? resume?.resumePayload : undefined,
           tracingContext: {
-            parentSpan: stepAISpan,
+            currentSpan: stepAISpan,
           },
           getInitData: () => stepResults?.input as any,
           getStepResult: (step: any) => {
@@ -1730,7 +1730,7 @@ export class InngestExecutionEngine extends DefaultExecutionEngine {
     writableStream?: WritableStream<ChunkType>;
     tracingContext?: TracingContext;
   }): Promise<StepResult<any, any, any, any>> {
-    const conditionalSpan = tracingContext?.parentSpan?.createChildSpan({
+    const conditionalSpan = tracingContext?.currentSpan?.createChildSpan({
       type: AISpanType.WORKFLOW_CONDITIONAL,
       name: `conditional: ${entry.conditions.length} conditions`,
       input: prevOutput,
@@ -1762,7 +1762,7 @@ export class InngestExecutionEngine extends DefaultExecutionEngine {
                 runCount: -1,
                 inputData: prevOutput,
                 tracingContext: {
-                  parentSpan: evalSpan,
+                  currentSpan: evalSpan,
                 },
                 getInitData: () => stepResults?.input as any,
                 getStepResult: (step: any) => {
@@ -1856,7 +1856,7 @@ export class InngestExecutionEngine extends DefaultExecutionEngine {
           runtimeContext,
           writableStream,
           tracingContext: {
-            parentSpan: conditionalSpan,
+            currentSpan: conditionalSpan,
           },
         }),
       ),